+++ src/main/groovy/org/gradle/api/plugins/cargo/CargoPluginTask.groovy
@@ -15,26 +15,33 @@
  */
 package org.gradle.api.plugins.cargo
 
-/**
- * Created by IntelliJ IDEA.
- * User: Ben
- * Date: 7/30/11
- * Time: 12:23 PM
- * To change this template use File | Settings | File Templates.
- */
 enum CargoPluginTask {
-    DEPLOY_REMOTE('cargoDeployRemote', 'Deploys WAR to remote container.'),
-    UNDEPLOY_REMOTE('cargoUndeployRemote', 'Undeploys WAR from remote container.'),
-    REDEPLOY_REMOTE('cargoRedeployRemote', 'Redeploys WAR to remote container.'),
-    RUN_LOCAL('cargoRunLocal', 'Starts the container, deploys WAR to it and wait for the user to press CTRL + C to stop.'),
-    START_LOCAL('cargoStartLocal', 'Starts the container, deploys WAR to it and then do other tasks (for example, execute tests).'),
-    STOP_LOCAL('cargoStopLocal', 'Stops local container.')
+    DEPLOY_REMOTE('cargoDeployRemote', 'Deploys WAR to remote container.', RemoteContainerTask, Action.DEPLOY),
+    UNDEPLOY_REMOTE('cargoUndeployRemote', 'Undeploys WAR from remote container.', RemoteContainerTask, Action.UNDEPLOY),
+    REDEPLOY_REMOTE('cargoRedeployRemote', 'Redeploys WAR to remote container.', RemoteContainerTask, Action.REDEPLOY),
+    RUN_LOCAL('cargoRunLocal', 'Starts the container, deploys WAR to it and wait for the user to press CTRL + C to stop.', LocalContainerTask, Action.RUN),
+    START_LOCAL('cargoStartLocal', 'Starts the container, deploys WAR to it and then do other tasks (for example, execute tests).', LocalContainerTask, Action.START),
+    STOP_LOCAL('cargoStopLocal', 'Stops local container.', LocalContainerTask, Action.STOP)
 
     final String name
     final String description
+    final Class taskClass
+    final Action action
 
-    private CargoPluginTask(String name, String description) {
+    private CargoPluginTask(String name, String description, Class taskClass, Action action) {
         this.name = name
         this.description = description
+        this.taskClass = taskClass
+        this.action = action
+    }
+
+    static enum Action {
+        DEPLOY('deploy'), UNDEPLOY('undeploy'), REDEPLOY('redeploy'), RUN('run'), START('start'), STOP('stop')
+
+        final String name
+
+        private Action(String name) {
+            this.name = name
+        }
     }
 }